move nucleotide meta-header to its sub-module folder and update alphabet.hpp
